송22
불리건바디,플레이어바디 구분,
실질적 히트박스가 되는 콜라이더 만듬,
임시로 싱글톤이라도 에너미가 플레이어 정보 받게 만들기

+++ CBullet.h
@@ -22,7 +22,7 @@ public:
 	CBullet(Vec2 _pos, RECT _rc, float _angle, float _speed,float _distance, int _height, CHARACTER _type, int size);
 	~CBullet();
 
-	virtual HRESULT init(float _angle, float _speed, vector2 _pt, float _height, float _distance, CHARACTER _type, int size);
+	virtual HRESULT init(float _angle, float _speed, vector2 _pt, float _height, float _distance, CHARACTER _type, float size);
 	virtual void release();
 	virtual void update();
 	virtual void render();
